+++ jjm/Navi-api/.env
@@ -19,3 +19,7 @@ FRONTEND_APP_URL=http://localhost:5173
 
 # 백엔드 주소
 APP_API_URL=http://localhost:8000
+
+# 구글
+# GOOGLE_CLIENT_ID=xxx.apps.googleusercontent.com
+# GOOGLE_CLIENT_SECRET=yyy
