+++ jjm/Navi-api/app.js
@@ -7,6 +7,9 @@ const passport = require('passport') // 인증 미들웨어
 require('dotenv').config() // 환경 변수 관리
 const cors = require('cors') // cors 미들웨어 -> ★api 서버는 반드시 설정해줘야 한다
 
+// swagger 추가
+const { swaggerUi, swaggerSpec } = require('./swagger')
+
 const indexRouter = require('./routes')
 const authRouter = require('./routes/auth')
 
@@ -48,12 +51,25 @@ const sessionMiddleware = session({
 })
 app.use(sessionMiddleware)
 
+// app.use('/auth', authRoutes) // 👈 꼭 필요!
+
+// swagger 추가
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec))
+
+// 구글 app.use 추가
+// app.use(
+//    session({
+//       secret: 'your_secret',
+//       resave: false,
+//       saveUninitialized: false,
+//    })
+// )
+
 app.use(passport.initialize())
 app.use(passport.session())
 
 app.use('/', indexRouter)
 app.use('/auth', authRouter)
-// app.use('/order', orderRouter)
 
 app.use((req, res, next) => {
    const error = new Error(`${req.method} ${req.url} 라우터가 없습니다.`)
