+++ jjm/Navi-api/package.json
@@ -22,9 +22,12 @@
       "multer": "^2.0.2",
       "mysql2": "^3.14.3",
       "passport": "^0.7.0",
+      "passport-google-oauth20": "^2.0.0",
       "passport-local": "^1.0.0",
       "sequelize": "^6.37.7",
-      "socket.io": "^4.8.1"
+      "socket.io": "^4.8.1",
+      "swagger-jsdoc": "^6.2.8",
+      "swagger-ui-express": "^5.0.1"
    },
    "devDependencies": {
       "nodemon": "^3.1.10"
