+++ jjm/Navi-api/passport/localStrategy.js
