+++ jjm/navi-frontend/src/styles/common.css
