+++ jjm/navi-frontend/src/styles/login.css
@@ -1,0 +1,33 @@
+/* 전체 배경 어둡게  */
+.overlay {
+   position: fixed;
+   top: 0;
+   left: 0;
+   width: 100vw;
+   height: 100vh;
+   display: flex;
+   justify-content: center;
+   align-items: center;
+   z-index: 1000;
+}
+
+/* 중앙 팝업 */
+.popup {
+   background-color: white;
+   padding: 30px;
+   border-radius: 12px;
+   box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
+   text-align: center;
+   min-width: 300px;
+}
+
+/* 닫기 버튼 */
+.close-btn {
+   margin-top: 20px;
+   padding: 8px 16px;
+   background-color: #ff5c5c;
+   color: white;
+   border: none;
+   border-radius: 6px;
+   cursor: pointer;
+}
