--- 전준모/navi-frontend/src/pages/SearchPwPage.jsx
@@ -1,12 +1,0 @@
-import SearchPw from '../components/auth/SearchPw'
-import { Container } from '@mui/material'
-
-function SearchPwPage() {
-   return (
-      <Container maxWidth="md" sx={{ marginTop: 10, marginBottom: 13 }}>
-         <SearchPw />
-      </Container>
-   )
-}
-
-export default SearchPwPage
